+++ api/src/Storage.ts
@@ -2,26 +2,32 @@ import { HexString } from '@polkadot/util/types';
 import { Option } from '@polkadot/types';
 import { u8aToU8a } from '@polkadot/util';
 
-import { ActiveProgram, IGearPages, ProgramMap } from './types';
-import { ProgramDoesNotExistError, ProgramExitedError, ProgramTerminatedError } from './errors';
+import { ActiveProgram, IGearPages, IProgram, PausedProgramBlockAndHash, PausedProgramMapValue } from './types';
+import {
+  PausedProgramDoesNotExistError,
+  ProgramDoesNotExistError,
+  ProgramExitedError,
+  ProgramTerminatedError,
+} from './errors';
 import { GearApi } from './GearApi';
 
 export class GearProgramStorage {
   constructor(protected _api: GearApi) {}
 
   /**
-   * Get program from chain
-   * @param programId
+   * ### Get program from chain
+   * @param id Program id
+   * @param at _(optional)_ Hash of block to query at
    * @returns
    */
-  async getProgram(programId: HexString): Promise<ActiveProgram> {
-    const programOption = (await this._api.query.gearProgram.programStorage(programId)) as Option<ProgramMap>;
+  async getProgram(id: HexString, at?: HexString): Promise<ActiveProgram> {
+    const programOption = (await this._api.query.gearProgram.programStorage(id, at)) as Option<IProgram>;
 
     if (programOption.isNone) {
-      throw new ProgramDoesNotExistError(programId);
+      throw new ProgramDoesNotExistError(id);
     }
 
-    const program = programOption.unwrap()[0];
+    const program = programOption.unwrap();
 
     if (program.isTerminated) throw new ProgramTerminatedError(program.asTerminated.toHex());
 
@@ -46,5 +52,29 @@ export class GearProgramStorage {
       );
     }
     return pages;
+  }
+
+  /**
+   * ### Get block number and hash of paused program
+   * @param id paused program id
+   * @param at _(optional)_ Hash of block to query at
+   * @returns
+   */
+  async getPausedProgramHashAndBlockNumber(id: HexString, at?: HexString): Promise<PausedProgramBlockAndHash> {
+    const storageOption = (await this._api.query.gearProgram.pausedProgramStorage(
+      id,
+      at,
+    )) as Option<PausedProgramMapValue>;
+
+    if (storageOption.isNone) {
+      throw new PausedProgramDoesNotExistError(id);
+    }
+
+    const storage = storageOption.unwrap();
+
+    return {
+      blockNumber: storage[0],
+      hash: storage[1],
+    };
   }
 }