+++ api/src/types/interfaces/events/program-changed.ts
@@ -1,15 +1,17 @@
-import { BlockNumber, H256 } from '@polkadot/types/interfaces';
 import { Enum, Null } from '@polkadot/types';
+import { BlockNumber } from '@polkadot/types/interfaces';
 
 export interface ProgramChangedKind extends Enum {
   isActive: boolean;
-  isInactive: boolean;
-  isPaused: boolean;
   asActive: { expiration: BlockNumber };
+  isInactive: boolean;
   asInactive: Null;
-  asPaused: {
-    code_hash: H256;
-    memory_hash: H256;
-    waitlist_hash: H256;
-  };
+  isPaused: boolean;
+  asPaused: Null;
+  isTerminated: boolean;
+  asTerminated: Null;
+  isExpirationChanged: boolean;
+  asExpirationChanged: { expiration: BlockNumber };
+  isProgramSet: boolean;
+  asProgramSet: { expiration: BlockNumber };
 }