+++ api/src/types/interfaces/program/index.ts
@@ -1,4 +1,4 @@
-export * from './base';
 export * from './pages';
 export * from './code';
 export * from './extrinsic';
+export * from './storage';