+++ idea/frontend/src/app/providers/events/Provider.tsx
@@ -1,8 +1,9 @@
 import { useState, useEffect } from 'react';
 import { useApi, ProviderProps } from '@gear-js/react-hooks';
 
-import { IdeaEvent, Section } from 'entities/explorer';
+import { EventRecords, IdeaEvent, Section } from 'entities/explorer';
 
+import { UnsubscribePromise } from '@polkadot/api/types';
 import { EventsContext } from './Context';
 
 const EventsProvider = ({ children }: ProviderProps) => {
@@ -10,7 +11,7 @@ const EventsProvider = ({ children }: ProviderProps) => {
   const [events, setEvents] = useState<IdeaEvent[]>();
 
   const subscribeToEvents = () =>
-    api.query.system.events(async (records) => {
+    api.query.system.events(async (records: EventRecords) => {
       const { createdAtHash } = records;
 
       if (createdAtHash) {
@@ -23,7 +24,7 @@ const EventsProvider = ({ children }: ProviderProps) => {
 
         setEvents((prevEvents) => (prevEvents ? [...newEvents, ...prevEvents] : newEvents));
       }
-    });
+    }) as unknown as UnsubscribePromise;
 
   useEffect(() => {
     if (!api) {