+++ utils/gear-hooks/src/utils/index.ts
@@ -20,4 +20,6 @@ const isLoggedIn = ({ address }: InjectedAccountWithMeta) => localStorage.accoun
 const getAutoGasLimit = ({ waited, min_limit }: GasInfo) =>
   waited ? min_limit.add(min_limit.mul(bnToBn(0.1))) : min_limit;
 
-export { getBalance, getAccount, isLoggedIn, getAutoGasLimit };
+const withoutCommas = (value: string) => value.replace(/,/g, '');
+
+export { getBalance, getAccount, isLoggedIn, getAutoGasLimit, withoutCommas };